+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64_x86_64-simulator/FitCloudKit.framework/Headers/FitCloudFirmwareVersionObject.h
@@ -296,8 +296,12 @@
 /// 是否支持显示气压
 @property(nonatomic, readonly) BOOL withBarometricPressure;
 
-/// 是否支持查询传音(Infinix)手表序列号(EUID, Equipment Unique Identifier)
-@property(nonatomic, readonly) BOOL allowQueryTranssionInfinixSerialNumber;
+/// 是否支持查询传音(Infinix)手表设备信息
+/// @note 包含以下信息:
+/// - 序列号(EUID, Equipment Unique Identifier)
+/// - 设备型号
+/// - 是否支持4G网络
+@property(nonatomic, readonly) BOOL allowQueryTranssionInfinixDeviceInfo;
 
 /// 是否支持切换耳机仓鼠标起始点预设
 @property(nonatomic, readonly) BOOL allowSwitchMouseStartPointPresetForEarphoneCase;