+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64_x86_64-simulator/FitCloudKit.framework/Headers/FitCloudKit+TranssionInfinix.h
@@ -56,12 +56,15 @@
 ///   - completion: the completion callback
 + (void)queryTranssionInfinixWatchMenstrualDataForMonth:(NSString *_Nonnull)month completion:(void (^_Nullable)(BOOL success, NSArray<NSNumber *> *_Nullable menstrualDataArray, NSError *_Nullable error))completion;
 
-#pragma mark - Query transsion infinix watch serial number
+#pragma mark - Query transsion infinix watch information
 
-/// Query Transsion Infinix Watch Serial Number
+/// Query Transsion Infinix Watch Information
 /// - Parameters:
 ///   - completion: the completion callback
-+ (void)queryTranssionInfinixWatchSerialNumberWithCompletion:(void (^_Nullable)(BOOL success, NSString *_Nullable sn, NSError *_Nullable error))completion;
+///     - deviceSn: the device serial number
+///     - deviceModel: the device model, e.g. `Infinix-XW4E`
+///     - isDevice4G: whether the device is 4G, `1` is 4G, `0` is not 4G
++ (void)queryTranssionInfinixWatchInformationWithCompletion:(void (^_Nullable)(BOOL success, NSString *_Nullable deviceSn, NSString *_Nullable deviceModel, NSNumber *_Nullable isDevice4G, NSError *_Nullable error))completion;
 
 @end