+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64_x86_64-simulator/FitCloudKit.framework/Headers/FitCloudModels.h
@@ -10,6 +10,6 @@
 #define FitCloudModels_h
 
 #import <FitCloudKit/FitCloudWorkoutSlot.h>
-
+#import <FitCloudKit/FitCloudAppSidePermissionStatusModel.h>
 
 #endif /* FitCloudModels_h */